Selection when reshaping view is now representative
The view was moved to far in the x and y direction compared to the
selection outline by exactly the size of the window border.

(Ammend to this and the previous patch, because apparantly I am blind
and did not see that the variable I was substracting was simply added to
the coordinates a few lines down. Was this added on purpose? Either way,
the selection being precise should be the correct way.)